Revert "Types (#197)"
This reverts commit cd02f5ec839ac06deaed64191688a15cfdb6c728.
