+++ frontend/components/AnalyticsDashboard.tsx
@@ -15,6 +15,17 @@ import _ from 'lodash';
 
 import { formatFileSize, formatNumber } from '@/lib/utils';
 
+interface ClusterSpecs {
+  cpus: number;
+  vendorId: string;
+  modelName: string;
+  cpuMhz: number;
+}
+
+export interface Tag {
+  type: string;
+  count: number;
+}
 interface AnalyticsDashboardProps {
   tokensProcessed: number; // This is actually bytes processed
   tags: Tag[];
